requirements/base.txt
bleach==1.4
django-braces==1.4.0
django-reversion==1.8.0
Django==1.6.5
docutils==0.11
psycopg2==2.5.2
pytz==2014.2
PyYAML==3.11
South==0.8.4
